+++ src/main/java/com/capgemini/start/domain/repository/GeneroRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.capgemini.start.domain.entity.Genero;
 
-public interface GeneroRepository extends JpaRepository<Genero, Integer> {
+public interface GeneroRepository extends JpaRepository<Genero, Long> {
 
 	boolean existsByDescricaoIgnoreCase(String descricao);
 
-	boolean existsByIdNotAndDescricaoIgnoreCase(Integer id, String descricao);
+	boolean existsByIdNotAndDescricaoIgnoreCase(Long id, String descricao);
 
 }
